+++ src/main/java/org/eclipse/keyple/card/calypso/CmdCardSelectFile.java
@@ -74,13 +74,11 @@ final class CmdCardSelectFile extends AbstractCardCommand {
         0x6700,
         new StatusProperties("Lc value not supported.", CardIllegalParameterException.class));
     m.put(0x6A82, new StatusProperties("File not found.", CardDataAccessException.class));
-    m.put(0x6119, new StatusProperties("Correct execution (ISO7816 T=0).", null));
+    m.put(0x6119, new StatusProperties("Correct execution (ISO7816 T=0)."));
     STATUS_TABLE = m;
   }
 
   private static final int TAG_PROPRIETARY_INFORMATION = 0x85;
-
-  private byte[] proprietaryInformation;
 
   /**
    * (package-private)<br>